checked working user can only access their own items

+++ backend/controllers/itemController.js
@@ -1,6 +1,7 @@
 //imports
 const asyncHandler = require("express-async-handler");
 const Item = require("..//models/itemModel"); //import Item model
+const User = require('../models/userModel'); //import User model
 
 // @desc    Get all items
 // @route   GET /api/items
@@ -22,6 +23,7 @@ const createItem = asyncHandler(async (req, res) => {
 
   const item = await Item.create({
     text: req.body.text,
+    user: req.user.id,
   }); //create item
 
   res.status(200).json(item);
@@ -33,9 +35,21 @@ const createItem = asyncHandler(async (req, res) => {
 const updateItem = asyncHandler(async (req, res) => {
   const item = await Item.findById(req.params.id); //find item by id
 
-  if (!item) {
+  if (!item) { //check if item exists
     res.status(400);
     throw new Error("Item not found");
+  }
+
+  //check for user
+  if(!req.user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  //check if user owns item
+  if(item.user.toString() !== req.user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
   }
 
   const updatedItem = await Item.findByIdAndUpdate(req.params.id, req.body, {
@@ -56,6 +70,18 @@ const deleteItem = asyncHandler(async (req, res) => {
   if (!item) {
     res.status(400);
     throw new Error("Item not found");
+  }
+
+  //check for user
+  if(!req.user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  //check if user owns item
+  if(item.user.toString() !== req.user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
   }
 
   await item.deleteOne(); //delete item
